requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
autobahn==23.6.2 ; python_version >= '3.9'
automat==22.10.0
cffi==1.16.0 ; python_version >= '3.8'
channels
daphne
constantly==23.10.4 ; python_version >= '3.8'
cryptography==41.0.5 ; python_version >= '3.7'
django==3.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1 ; python_version >= '3.5'
jsonschema==4.20.0 ; python_version >= '3.8'
jsonschema-specifications==2023.11.1 ; python_version >= '3.8'
msgpack
pyasn1==0.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyopenssl==23.3.0
pytz==2023.3.post1
pyyaml==6.0.1 ; python_version >= '3.6'
redis==5.0.1 ; python_version >= '3.7'
referencing==0.31.0 ; python_version >= '3.8'
rpds-py==0.13.1 ; python_version >= '3.8'
service-identity==23.1.0
setuptools==69.0.2 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
twisted[tls]==23.10.0 ; python_full_version >= '3.8.0'
txaio==23.1.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
uritemplate==4.1.1 ; python_version >= '3.6'
zope-interface==6.1 ; python_version >= '3.7'
